dtype

+++ oneat/NEATModels/neat_diamondstandard.py
@@ -272,7 +272,7 @@ class NEATEynamic(object):
         return self.marker_tree
     
     def predict(self, imagename,  savedir, n_tiles=(1, 1, 1), overlap_percent=0.8,
-                event_threshold=0.5, event_confidence = 0.5, iou_threshold=0.1,  dtype = 'uint16',
+                event_threshold=0.5, event_confidence = 0.5, iou_threshold=0.1,  dtype = 'uint8',
                 marker_tree = None, remove_markers = False, normalize = True,  nms_function = 'iou'):
 
 
@@ -319,7 +319,7 @@ class NEATEynamic(object):
            self.default_pass_predict() 
 
         if self.normalize: 
-              self.image = normalizeFloatZeroOne(self.image, 1, 99.8)   
+              self.image = normalizeFloatZeroOne(self.image, 1, 99.8, dtype = dtype)   
 
     def default_pass_predict(self):
         eventboxes = []
